Changes to make compilation work in Gingerbread. Must update build/
Removed tags where not needed and change some to optional.
- static libraries do not need LOCAL_MODULE_TAGS set (defaults
to optional) and will get built and included where needed.
- shared libraries libamavjni and libamplayer have their tag
set to optional and are not used anywhere so they have not been
included in the PRODUCT_PACKAGES variable.
- amlogic.libplayer.xml has been set to optional and is included
in either build/target/product/generic.mk or device/amlogic/ by
adding them to PRODUCT_PACKAGES.